ADDED in additional classes and methods
Added in additional logic classes to address: Separation of Concerns,
reusability, testability, scalability and flexibility. Also added in
a Location class. Wasn't entirely sure how to model the interaction
between logic classes and other classes so i just used the dashed arrows
with arrows pointing in both directions. Can update this later.

Added in a number of methods as a rough placeholder but anticipate
they will change. They are more intended as a start point than anything
concrete.

This class diagram is not intended to be final just more added to the
totality of the task.